chore: run scripts with tsx
